+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2019-09-11
+ - Additional colour picking shortcuts with copy hex: CTRL-C and copy as code: CTRL-SHIFT-C.
+ - RenderTexture colour picking now updates as it should.
+
 ## [1.0.4] - 2019-09-10
  - Added support for CustomRenderTexture.
  - Fixed colour picking with RenderTextures.